styles teacher and student with background color and vertical border to simplify design
chose to use absolute positioning for the left section because then top and bottom 0 can be used to create a vertical border that expands the page. Also, a fixed width is given to the login form with text-align right, which prevents the page from being responsive as it stands. Maybe fix later, for now it's fine.