adds error messages and contextual submit button to student responses
Error Message Compromise: because the form is modifying the user object, but the attributes being filled are from the Response model, the target.errors.full_messages => "Responses is invalid," rather than the much more useful "This question is missing."

Therefore the compromise is just counting the number of incorrect responses. Later, consider using client-side validation and figuring out how to fill up the form with any completed responses.

Conextual Submit Button: used a partial to show 3 buttons: next set of questions, complete survey, and survey is already complete. This is necessary because students are not persisted in the database, meaning anybody could see a finished survey page by the student number.

Consider changing to funny urls, or better yet, rolling the student model as a role within the user model (but that requires some hardcore changes).