adds nav for sub factors and persists logo for users to enable simple navigation
Preference is to have fixed position nav at the top. However, it takes longer to do with JQuery, and in the interest of "making it work" before "making it right," I simplified to a logo on the top left that links to user show page, and links at the bottom for subfactors.