adds incomplete tracker to subfactor calculation to stop breakage
A little messy to continuously track incomplete and subtract from calculation for the four methods used to calculate the graphs (2 Avg Scores, and 2 Class Learning Preferences methods). Consider an alternate solution, like adding an additional column in the database that tracks incomplete students and omits them from calculation. Probably not worth it though.